some more tests for utils.maybe_await
